Implementish work spec metadata, deletion.
This adds RPC calls ControlWorkSpec, GetWorkSpecMeta, Clear, and
DelWorkSpec.  There are Go-side API calls to modify the work spec metadata,
including details like "is it paused", "is it continuous", "what is its
weight", etc., but the rest of the system does not use these right now.

Implementing Clear() means that the "run the Python tests against this"
recipe can be updated so that successive tests don't step on each other.