Switch Namespace.WorkSpec() to return an error on not-found.
In particular it returns an instance of coordinate.ErrNoSuchWorkSpec.
This seems to be more in line with standard Go practice.  It certainly
simplifies the jobserver implementation, where we can now treat
"no work spec" the same way as any other API-returned error instead of
having to special-case the "didn't return an error but also didn't return
a work spec" case.

Speaking of, since the CBOR-RPC interface has both errors and exceptions,
remap the ErrNoSuchWorkSpec "exception" to a string error, since every
call treats it this way.