Switch add_work_units to mapstructure, remove cleanup cruft.
Change the marshaling that happens to convert add_work_units's
list-of-tuples to list-of-struct to work by changing the tuple to a
string-keyed map and then invoking mapstructure, because a library
someone else has written to do type marshaling is better than writing
our own.

Speaking of, this was the last use of several of the cborrpc.cleanup
marshaling functions (SloppyString, StringKeyedMap, ...) so remove those.