Port one complete Python jobserver test.
This is test_data_updates() from coordinate/tests/test_job_client.py.
The test runs through a complete sequence (create a work spec, add a
work unit, get a work unit, complete the work unit) with a fair amount
of intermediate checking, so it's good to have it natively.

There are also a bunch of building blocks here that may want to be
refactored, but having written out the half-dozen lines per call (to
turn an object into the correct Go type, get the correct Go object set
back, and validate the results) it's at least worth having to
copy-and-paste.