Add jobserver "get many" tests, fix for postgres
Port two tests that get multiple work units, the second limiting it with
a work spec max_getwork option.

Implement this for the postgres backend by pulling every work spec's
complete metadata, including work unit counts.  This is clearly going
to be the input to the scheduler when it appears.  (We may need to
add some indexing to make it work efficiently.)  Start building a
set of constants for various SQL names.