Add a work spec scheduler.
This works based on work specs' "priority" and "weight" fields.
Priority is absolute, within a priority level weight controls the number
of expected pending jobs and the scheduler tries to assign a work spec
so the ratio of pending work units matches the ratio of the weights.

This scheduler does not match the "traditional" Python scheduler, though
there is a placeholder for it that documents its behavior.  This does
enable passing another ported test 99.3% of the time, though.

Also add (because the scheduling takes it into account) a knob for
the amount of memory available to the worker and the matching min_gb
constraint (which, again, some ported tests will depend on).