Eradicate the (unported) "compatible" scheduler.
Python coordinate has a somewhat complicated work spec scheduler that
tries to look at the chains of what work specs trigger which other work
specs, and gives precedence to the most-downstream work specs.  In practice
this is much more complicated than is justified, exotic work spec flow graphs
are rare, and if there is a long queue going into part A generating
prioritized work units in part B then we don't want to blindly run everything
in part B immediately.

Remove all mentions and support for this, beyond the top-level README.
Don't extract the "then_preempts" field from a work spec; this only affected
that scheduler (and the aforementioned use case where we want to keep running
the earlier part to prioritize the late part).

There is a newer, much simpler scheduler, which should achieve a practically
correct behavior, it is much easier to explain, and (since it is probabalistic
in many interesting cases) it should not starve out very-low-weight work specs
(if something has weight=1 it should run at least sometimes).