Postgres: minimize repeated work unit counting.
If there are concurrent calls to RequestAttempts() that start to think
they'll do the same work unit, the database will force all but one of them
to start over (the SELECT...FOR UPDATE statement will fail and the withTx()
wrapper code will retry).  When this happens, it's likely that the work unit
counts haven't changed much (and the scheduler decision is the same), but
fetching the counts is really expensive.  Just try to get work units again,
for the same work spec, if this does happen.

The flip side of this is that it's possible the "other" worker picked up
the very last work units, which would result in this returning no attempts.
If *that* happens then we do need to go back and re-run the scheduler.

At any rate, on my tests this is good for about a 3x performance gain on
the slowest concurrent tests.