Accept more things in a work unit "output".
It can be a dictionary (of keys to data dictionary), or a tuple or a slice
(of anything).  If the latter, the individual values can be a string, a
[]byte (both produce that key and empty data), or a tuple or a slice;
and if the last, the tuple is of (key, data, metadata).  A mixed list of
strings and tuples appears in the wild (in streamcorpus-pipeline).

Probably not coincidentally, this is almost the exact same set of things
the CBOR-RPC add_work_units call expects, so move that code from jobserver
to coordinate core.  That also slightly expands the list of things that's
acceptable to add_work_units (which can now take bare strings as well).