Fix chaining when a Renew() added the output.
In the case where a work unit was created with no output in its data,
it was subsequently renewed with updated data that included an output key,
and then finished with nil data, both the memory and postgres backends
saw Finish(nil) and jumped back to the original unit data, ignoring the
updated data from the Renew() call.  Fix that.