Postgres: allMetas(true) work unit count performance.
SELECT COUNT(*) has to manually walk through all of the things.  If
this is the number of available work units, that count can be large.  But,
in practice, it only matters (for the internal allMetas()) that a work
spec has one or more available units.  We do need the actual pending
count.

Some strategic EXPLAINing suggests that performance should be much better
(with thousands of available units) if we separately ask the database
for pending counts and whether there is at least one available work unit;
do that.