Fix a postgres bug with duplicate "output".
If work spec "one" declares "then": "two", and its work units generate
"output" fields with duplicate work unit keys, don't fail on the second
one.  Instead reuse the spec.addWorkUnits() helper that does this exact
thing correctly.  (Which implies getting a workSpec object for the next
work spec in the chaining code, which implies refactoring that code so
that it can be called for an existing transaction.)