jobserver: allow UpdateWorkUnit(AVAILABLE).
There is a race condition in the Python worker code where a parent could
start trying to kill off soon-to-expire work units, but since the system
is busy, it doesn't get to one before it actually does expire.  This results
in attempting an AVAILABLE (expired) to FAILED transition, which jobserver
reports as an error, which the Python worker translates to an exception,
which, since it's in the parent, explodes the worker.

Instead allow most actions on available work units.  You can use
UpdateWorkUnit to change a work unit's data, or directly transition to
other states.  Internally this creates an attempt based on the worker ID
passed in the request, then steps that immediately to finished/failed.
(Or you could use UpdateWorkUnit(status=PENDING) to forcibly start a specific
work unit; though I don't think any Python code does this.)