REST client and server.
This commit makes API changes:

* WorkSpec.WorkUnit() returns a new error, ErrNoSuchWorkUnit, if the
  requested work unit does not exist.  It was previously specified to
  return nil and no error.

The coordinated daemon runs an HTTP REST server on a specified port.
Once this is running, the restclient package provides a coordinate
implementation that talks to this.  A Coordinate relay server could
be set up using -backend http://...

The most prominent limitation of the REST system is that Attempts
cannot be uniquely identified, so the same worker cannot attempt the
same work unit more than once per second.  In typical use this should
never come up.

Changes to common tests for this were small but necessary, mostly
handling cases where a zero time went through a network round-trip and
was no longer exactly zero, and advancing the test clock to avoid the
attempt limitation above.