postgres: lock the work unit table in addWorkUnit()
This avoids a concurrency issue where two things both try to add the
same work unit.  Both SELECT ... FOR UPDATE an existing unit, and neither
finds it, so both go on to execute an INSERT for it, and one hits a
constraint violation.

It's something of a performance hit, though I haven't quantified it.
(It looks like it should only stall other things that modify the work
units table, but not reads or other row locks.)  See comments for
alternatives.