Fix URL parsing bug.
The issue here is that `url` is used as a variable name in addition
to being the namespace for the imported package `net/url`. Normally
this would return a compile error, but the `*URL` type (the type of
variable `url`) has a `Parse` method *and* there is also a package
level function called `Parse` with the same type signature. The end
result is that `url.Parse(...)` is actually `((*URL)(nil)).Parse(...)`,
which panics on a nil error.

Also, add a regression test.