Add the ability to run a work unit in the future.
The WorkSpec.AddWorkUnit() call changes to take a new WorkUnitMeta structure
instead of a single priority parameter.  This includes both the priority and
the earliest time the unit can be allowed to run (could be a zero time).
In turn, this introduces a new work unit status DelayedUnit.  If using
chained work specs, the "output" field can be lists of 3-item lists, where
the third item is a metadata dictionary; passing "delay": 60 in that dictionary
causes the newly created work unit to be delayed by 60 seconds.

There are corresponding (immediate, breaking) changes in the REST API; of
note you cannot directly change only a work unit's priority without also
potentially changing its not-before time.  (I think the API is new enough I
can get away with that.)

The WorkUnit.Priority() and .SetPriority() calls are now marked as deprecated,
since .Meta() and .SetMeta() can do the same thing, but they are not
immediately deleted.

There are significant internal changes in the PostgreSQL implementation: all
of the SQL support code got pulled into a new postgres/sql.go file, which
includes a helper type for the ($1, $2, ...) parameter list.  This helps
clarify some of the complex queries (which had comments making strong
assumptions about what each numbered parameter is).  This was more or less
necessary for adding a not_before column to the work_units table, since the
"is this work unit available" query also now needs to take into account "is
it delayed" which in turn needs the current (application-controlled) time as
a parameter.