postgres: more SQL cleanups
Finish tweaking constants.go so that fixed query parameter numbers ($1) do
not appear anywhere; anything that refers to a query param does so via the
queryParams list object.  Make the naming a little more consistent too,
so "fooIsThisBar" is a constant from an SQL join and "fooIsBar(&qp, bar.id)"
is a query param version.  This in particular changes the meaning of
workUnitInSpec from the previous commit (it was the join constant and is now
workUnitInThisSpec, it replaces what was workUnitInOtherSpec() which looked
odd when it was the only WHERE clause part).