postgres: uninteresting SQL rewrites.
I wanted to make the default transaction mode SERIALIZABLE, and use that
to avoid the table lock in addWorkUnits(), but this doesn't seem to work
as advertised.  Nevertheless, force every operation into a transaction, and
add a flag to withTx() to allow read-only transactions.  Add a couple more
helpers for common cases and turn a couple more queries into build*() or
fieldList forms.