postgres: eliminate the global lock in addWorkUnits().
This was flagged in code as a bad idea, rightfully, and seems to in fact
be a bottleneck in reality.  The fundamental problem is that, if the work
unit already exists, addWorkUnits() is defined to update it, which is an
"upsert" operation; Google "why upsert is hard".

There are three cases when addWorkUnits() gets called.  When it gets called
creating a continuous work unit to immediately create an attempt, the work
unit is expected to not exist and there is a higher-level retry loop.  The
other cases (single spec AddWorkUnit() and handling "outputs" from chained
specs) require consistency but not necessarily atomicity _per se_.  Also,
the common case for addWorkUnits() should normally be that the unit *doesn't*
exist.

So, split up addWorkUnits() into a separate "insert" part (for continuous)
units, and otherwise run insert, update, ... in a client-side loop, trapping
constraint-violation errors on the insert and noticing when the update doesn't
update anything.

On the benchmarks in this package on my limited setup, serial
WorkUnitCreation is twice as fast and ConcurrentWorkUnitCreation 3x better.
ConcurrentUnitOutput also does less work and improves a little;
ConcurrentMultiAttempts is a little slower.