postgres: Performance improvements in DeleteWorkUnits.
Creating an index on attempt(work_unit_id) massively improves the behavor of
DELETE FROM work_unit ..., because for every work unit, it needs to find
corresponding rows in the attempt table.

If you're trying to DeleteWorkUnits while other parties are trying to
RequestAttempts, have DeleteWorkUnits run in batches of 1000 work units at a time,
and trap a 40P01 ("deadlock") PostgreSQL error.