postgres performance: flatten RequestAttempts queries
Previously we would make a SELECT query to get work units; for each INSERT a
single attempt; and for each UPDATE to set the active attempt ID.  Flatten those
into a single, complicated, WITH-based statement that SELECTs all the work
units, does a single batch INSERT to create the attempts, and then a single
batch UPDATE to set the active attempt IDs.  This is good for roughly 30% on
"coordbench do" with 30K work units.