postgres: SET TRANSACTION ISOLATION LEVEL globally
Previously we set it in every call to withTx(); but it's also possible to
set it in a connection parameter (and github.com/lib/pq has simplified
syntax to set it).  This seems to be good for an extremely modest performance
improvement, which isn't surprising, since there are fewer statements
executed but the statements that go away aren't very expensive.

As a side effect of this, remove any effect from declaring transactions
read-only (it doesn't obviously improve performance to do so).