postgres: performance and correctness of continuous units
When the scheduler returns a work spec, we already know whether or not
it has work units, so don't run the (expensive) query to try to create
attempts when we expect it to return nothing.

When we do create continuous units, the first thing we do is SELECT ...
FOR UPDATE the work spec, but while we hold that update lock we never do
any checking that creating the work unit is correct.  As a simple and
lightweight check, validate that the next-continuous time is still valid.
This can avoid a problem where multiple clients all pick the same
continuous work spec, all see it has no available work units, and all
decide to create a continuous unit, serializing on the update lock but
all ultimately doing it.