postgres: only run one expiry at a time
There is a global query to clear the active attempt from work units when
that active attempt expires, which turns out to be surprisingly expensive
if the system is busy.  Arrange for only one of those to be running at a
time (via sync.Cond); if someone else wants expiry to have happened, they
just block until the instance of it that is already running has finished.

This arrangement should get the practical benefits of running expiry
frequently but also pass tests that depend on it having run before calls
like RequestAttempts.