Switch tests from "gocheck" to "testify".
This brings a couple of advantages.  Since testify uses a normal testing.T
setup, it doesn't have an alternate set of command-line flags to remember.
Being able to use test assertions in conditionals is quite useful.  And
this change means that we do not need to reinvent parallel test execution
for benchmarks, or have separate parallel and sequential versions -- normal
'go test -bench . -cpu N' does that on its own.