postgres: specifically track the work spec in attempt table.
Scheduling happens very often, and it needs to know "find the number of
pending attempts in each work spec".  Before this change, this query was
prone to happening by scanning the list of every work unit, and pairing
it up with attempts (and I couldn't find a clever SQL rewrite that would
avoid this), so this makes routine execution a lot quicker.

This can lead to small differences in scheduling if there are pending
attempts that are not the active attempts for their work unit.  I can
think of two ways for this to happen: explicit calls to ClearActiveAttempt()
and explicit calls to MakeAttempt() that overwrite an existing attempt.
Both of these are unusual in routine operation.  Specific configurations
could have wider-reaching affects: if a work spec has max_getwork=1, and
an attempt is "lost" (e.g. its worker crashes), even if you
ClearActiveAttempt() on it, the scheduler will consider it "busy" and
not create a new attempt until the old attempt expires.