postgres: simplify continuous unit flow a little
Instead of SELECT work_spec FOR UPDATE, creating the continuous unit, and
then UPDATE the work_spec next-continuous time, just do one shot of UPDATE
the work_spec table first, testing the next-continuous time server-side.
This still causes the server to bounce concurrent continuous work unit
creation, sometimes, but also causes it to not block things on that row
lock.  This seems to improve things a little when there is only continuous
work to do.