PostgreSQL: use an advisory lock per work spec when getting work.
This reduces load on the database if one work spec has most of the work,
and there are many workers.  What would happen before is that all of the
workers would try to get work, and most would hit a retryable concurrency
failure in the database.  Instead, have them block at the point where they've
decided what work spec they're going to work on.

(Explicit database locks feel a little hokey to me.  A better solution
might be to launch a goroutine per work spec specifically to do
RequestAttempts, but this needs some significant work to actually happen.)