Postgres: work_unit_ordering index only has available units.
Change the work_unit_ordering index to only include work units WHERE
active_attempt_id IS NULL, which includes available (and delayed) work units
but not any pending or completed ones.  This addresses an issue at scale where
trying to do work correctly finds the index, but all of the highest-priority
units have already been done, so they need to be scanned through before the
next-highest-priority units happen.

While we're here and looking at indexes, drop the work_unit_spec_attempt index.
I think it was a try to get an index-only scan for one particular case, which
isn't a big performance difference when it happens.  In practice it's not used
(definitely not in the do-work critical path) and both constituent values are
indexed independently.