Postgres: fix continuous specs with intervals > 1 day
The SQL duration parsing code didn't actually understand Postgres's
syntax for this; fixed it.  (It is different from the official SQL
syntax and from the syntax that you write durations in.)

The issue is actually kind of annoying if you hit it: having a work
spec that declares "interval": 86400 (or longer) causes all calls to
RequestAttempts to fail, deleting it or setting a shorter interval should
work around it.