Fix a REST server crash on network hangup.
The restserver code was trying to write directly to the wire, and
panic()king on most errors.  If, in particular, there was a network
issue and all writes to the response writer fail, then MustEncode() on
the primary response will panic, and the recovery would try to
MustEncode() an error, which also fails, and that final panic would
trickle up to the point where it killed the server.

The solution to this is to encode to an intermediate buffer (these
responses should not be huge usually), then try to write the contents
of that.  Use the codec library's Encode(), which returns errors, not
MustEncode(), which panics.  There is a little bit of refactoring to
share the "encode then write out" sequence.