postgres: duplicate AddWorkUnit would bleed over into other work specs
If it went to UPDATE the work unit that already existed instead of
INSERTing a new one, the UPDATE statement wasn't limited to the single
work spec.  This caused both a consistency problem (if the same work unit
key was in multiple work specs, it would have unintended side effects) and
a performance problem (there is not an index for "find one work unit by
name" but there is for "find one work unit by work spec and name", so the
buggy code was a sequential scan over all work units).