postgres: use internal types in internal attempt creation code.
Consistently use *attempt, not the coordinate.Attempt interface, inside
the guts of attempt creation.  This avoids a type cast (where we statically
know the type) when the max-retries logic needs to fail a newly created
attempt inside the current transaction, at the cost of copying the result
array at the outside of Worker.MakeAttempts.