postgres: improve throughput when max_attempts checking is on.
The postgres RequestAttempts() code gets an advisory lock, limiting database
churn when there are multiple workers trying to get work for the same work
spec.  The previous code checked whether any newly-created attempts are
for work units with excessive attempts inside the transaction holding this
lock.  Change this to exit the transaction, positively acquiring the attempts,
and _then_ maybe expire them, allowing other workers to continue getting
work while we're doing this checking.

As a consequence, change the behavior of RequestAttempts() so that it will
return fewer than max_getwork attempts if some of them are expired, rather
than trying to refill the queue (with smaller database batches) trying to
get all of the attempts we could.