worker: fix data race
This commit fixes a data race where the worker's cancellation map was
read and written to simultaneously from multiple goroutines.

We fix this by using a sync.Map. While I think a regular map with a mutex
would be better for type safety, this was an easier change to make.
In particular, it's not clear if there would be any performance regressions
from a mutex (or deadlocks from a misapplied mutex), so we go the safe
route and use a safe concurrent map.