Adds NumAttempts() method
NumAttempts() should be heavily preferred to len(Attempts()), as
the latter will issue a network request for every attempt. When
there are items that have hundreds of attempts this can slow things
down significantly.

Initially I had intended to just add a NumAttempts() method to the
client, but it turns out this is the same interface used by
coordinate internally. I added this method to all other implementors,
but note that the REST client is still calling the old /attempts
endpoint and then returning the length.

Ideally we would call the new endpoint but in the interest of changing
as little as possible I opted for the path that doesn't require
redeploying the server.